fix: email service
